requirements.txt
mcp[cli]>=1.0.0
google-generativeai>=0.5.0
httpx>=0.27.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
webdriver-manager>=4.0.0
pandas>=2.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
aiohttp>=3.9.0
fake-useragent>=1.4.0
lxml>=5.0.0
